[lsp] connectors should delegate to singleton language client
Fixes #19
