+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sprotty-theia",
-  "version": "0.3.0-next.0",
+  "version": "0.2.0",
   "description": "Glue code for Sprotty diagrams in a Theia IDE",
   "license": "(EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0)",
   "keywords": [
