+++ src/sprotty/sprotty-language-client-extension.ts
@@ -1,0 +1,103 @@
+/********************************************************************************
+ * Copyright (c) 2018 TypeFox and others.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License v. 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * This Source Code may also be made available under the following Secondary
+ * Licenses when the conditions for such availability set forth in the Eclipse
+ * Public License v. 2.0 are satisfied: GNU General Public License, version 2
+ * with the GNU Classpath Exception which is available at
+ * https://www.gnu.org/software/classpath/license.html.
+ *
+ * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
+ ********************************************************************************/
+
+import URI from '@theia/core/lib/common/uri';
+import { EditorManager } from '@theia/editor/lib/browser';
+import { ILanguageClient, LanguageClientContribution, Location, NotificationType } from '@theia/languages/lib/browser';
+import { injectable } from 'inversify';
+import { ActionMessage } from 'sprotty/lib';
+
+export interface ActionMessageReceiver {
+    receivedThroughLsp(message: ActionMessage): void
+}
+
+export interface OpenInTextEditorMessage {
+    location: Location
+    forceOpen: boolean
+}
+
+const acceptMessageType = new NotificationType<ActionMessage, void>('diagram/accept')
+const didCloseMessageType = new NotificationType<string, void>('diagram/didClose')
+const openInTextEditorMessageType = new NotificationType<OpenInTextEditorMessage, void>('diagram/openInTextEditor')
+
+@injectable()
+export class DiagramLanguageClient {
+
+    actionMessageReceivers: ActionMessageReceiver[] = []
+
+    constructor(readonly languageClientContribution: LanguageClientContribution,
+                readonly editorManager: EditorManager) {
+        this.languageClientContribution.languageClient.then(
+            lc => {
+                lc.onNotification(acceptMessageType, this.receivedThroughLsp.bind(this))
+                lc.onNotification(openInTextEditorMessageType, this.openInTextEditor.bind(this))
+            }
+        ).catch(
+            err => console.error(err)
+        )
+    }
+
+    openInTextEditor(message: OpenInTextEditorMessage) {
+        const uri = new URI(message.location.uri)
+        if (!message.forceOpen) {
+            this.editorManager.all.forEach(editorWidget => {
+                const currentTextEditor = editorWidget.editor
+                if (editorWidget.isVisible && uri.toString(true) === currentTextEditor.uri.toString(true)) {
+                    currentTextEditor.cursor = message.location.range.start
+                    currentTextEditor.revealRange(message.location.range)
+                    currentTextEditor.selection = message.location.range
+                }
+            })
+        } else {
+            this.editorManager.open(uri).then(
+                editorWidget => {
+                    const editor = editorWidget.editor
+                    editor.cursor = message.location.range.start
+                    editor.revealRange(message.location.range)
+                    editor.selection = message.location.range
+                })
+        }
+    }
+
+    sendThroughLsp(message: ActionMessage) {
+        this.languageClientContribution.languageClient.then(lc => lc.sendNotification(acceptMessageType, message))
+    }
+
+    receivedThroughLsp(message: ActionMessage) {
+        this.actionMessageReceivers.forEach(client => {
+            client.receivedThroughLsp(message)
+        })
+    }
+
+    get languageClient(): Promise<ILanguageClient> {
+        return this.languageClientContribution.languageClient
+    }
+
+    didClose(clientId: string) {
+        this.languageClientContribution.languageClient.then(lc => lc.sendNotification(didCloseMessageType, clientId))
+    }
+
+    connect(client: ActionMessageReceiver) {
+        this.actionMessageReceivers.push(client)
+    }
+
+    disconnect(client: ActionMessageReceiver) {
+        const index = this.actionMessageReceivers.indexOf(client)
+        if (index >= 0) {
+            this.actionMessageReceivers.splice(index)
+        }
+    }
+}
